cyber-funk-marker.js: initial commit.
This component can be placed on any entity that should be tracking a marker. By supplying it with the id of the marker entity in the "constructor" (cyber-funk-marker="ADD_MARKER_ID_HERE") it will identify it and track its position. On top of that, a time until invisible delay can be provided.